requirements.txt
# Core Backend
fastapi
uvicorn[standard]
pydantic
python-dotenv
# PDF Processing
pdfplumber
pdf2image
pytesseract
python-docx
Pillow
# AI / ML
google-generativeai
groq
openai
scikit-learn
# Web Scraping
firecrawl-py
requests
playwright
# HTTP (must be modern for Python 3.13)
httpx>=0.27